POSTのタイミングで他のユーザーの位置情報もGET

+++ Mobile/app/src/main/java/com/example/hideandseek/data/datasource/remote/ResponseData.kt
@@ -12,9 +12,9 @@ class ResponseData {
     )
 
     data class ResponseGetSpacetime(
-        @Json(name = "latitude") val latitude: Float,
-        @Json(name = "longitude") val longitude: Float,
-        @Json(name = "altitude") val altitude: Float,
+        @Json(name = "latitude") val latitude: Double,
+        @Json(name = "longitude") val longitude: Double,
+        @Json(name = "altitude") val altitude: Double,
         @Json(name = "objId") val objId: Int
     )
 }
